refactor: move pty to node pty to make it easier to replace node pty with another module
